#82 [RENAME] 웹소켓 관련 파일 및 폴더 이름 변경
